feat: stop users from accessing payment if already paid
